requirements.txt
Flask
Flask-Cors
PyJWT
SQLAlchemy
requests
pytest
passlib==1.7.4
bcrypt==4.0.1
pytest-check
coverage
pytest-cov
